[1/2] Frameworks: Add per app controls for keyguard notifications
Nice done by google but the UX is a problem especially for ppl who are using a lot apps and just want
to see from important apps the notifications on the lockscreen.

This commit adds the ability to

- enable/disable per app the keyguard notification at all
- enable/disable per app ongoing notifications on the keyguard

We handle this over the app policy conf file like the other per app notification options.

Initial LP implementation by Lars Greiss, ported to MM by Yank555.lu

Signed-off-by: Jean-Pierre Rasquin <yank555.lu@gmail.com>

 Per app keyguard notification fix (https://github.com/beanstown106)

*fixed settings not persisting through a reboot
*cleaned up code a little bit
*added missing bits that were missed in the initial port to marshmallow
*also added missing enforceSystemOrSystemUI bit from lp